+++ bin/workflow_glue/tests/conftest.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+"""Pytests argument definitions."""
+
+
+def pytest_addoption(parser):
+    """Define command line arguments for pytest."""
+    parser.addoption(
+        "--test_data",
+        action="store",
+        default="/host/test_data"
+    )